requirements.txt
textual>=0.40.0
pyyaml>=6.0
